fix: added validations , error handling and simmilar logical pacthes between employee and department

+++ src/controller/employee.contoller.ts
@@ -8,8 +8,11 @@ import { CreateAddressDto, UpdateAddressDto } from "../dto/address.dto";
 import errorsToJson from "../../errorstojson";
 import { RequestWithUser } from "../utils/requestWithUser";
 import authorize from "../middleware/authorize.middleware";
-import { AnyBulkWriteOperation } from "typeorm";
 import { Role } from "../utils/role.enum";
+import Department from "../entity/department.entity";
+import dataSource from "../db/data-source.db";
+import EntityNotFoundException from "../exceptions/entityNotFound.exception";
+import { ErrorCodes } from "../utils/error.code";
 
 class EmployeeController {
   public router: express.Router;
@@ -19,11 +22,11 @@ class EmployeeController {
 
     this.router.get("/", this.getAllEmployees);
     this.router.get("/:id", this.getEmployeeById);
-    this.router.put("/:id", this.updateEmployee);
-    this.router.patch("/:id", this.updateEmployee);
-    this.router.delete("/:id", this.deleteEmployee);
-    this.router.post("/login", this.loginEmployee);
     this.router.post("/", authorize, this.createEmployee);
+    this.router.put("/:id", authorize, this.updateEmployee);
+    this.router.patch("/:id", authorize, this.updateEmployee);
+    this.router.delete("/:id", authorize, this.deleteEmployee);
+    this.router.post("/login", this.loginEmployee);
   }
 
   public getAllEmployees = async (
@@ -44,11 +47,7 @@ class EmployeeController {
       const employee = await this.employeeService.getEmployeeById(employeeId);
 
       if (!employee) {
-        const error = new HttpException(
-          404,
-          `No employee found with id: ${req.params.id}`
-        );
-        throw error;
+        throw ErrorCodes.EMPLOYEE_WITH_ID_NOT_FOUND;
       }
 
       res.status(200).send(employee);
@@ -66,24 +65,37 @@ class EmployeeController {
     try {
       const role = req.role;
       if (role != Role.HR) {
-        throw new HttpException(
-          403,
-          "You are not authorized to create employee"
-        );
+        throw ErrorCodes.UNAUTHORIZED;
       }
+
       const employeeDto = plainToInstance(CreateEmployeeDto, req.body);
+      console.log(employeeDto);
       const errors = await validate(employeeDto);
       if (errors.length) {
         console.log(errorsToJson(errors));
         throw new HttpException(400, JSON.stringify(errors));
       }
+
+      const departmentRepository = dataSource.getRepository(Department);
+      const departmentEntity = await departmentRepository.findOneBy({
+        name: employeeDto.department.name,
+      });
+
+      if (!departmentEntity) {
+        throw new EntityNotFoundException({
+          CODE: "DEPARTMENT_NOT_FOUND",
+          MESSAGE: "Department not found",
+        });
+      }
+
       const newEmployee = await this.employeeService.createEmployee(
         employeeDto.name,
         employeeDto.email,
         employeeDto.age,
         employeeDto.password,
         employeeDto.role,
-        employeeDto.address
+        employeeDto.address,
+        departmentEntity
       );
       res.status(201).send(newEmployee);
     } catch (err) {
@@ -93,33 +105,53 @@ class EmployeeController {
   };
 
   public updateEmployee = async (
-    req: express.Request,
+    req: RequestWithUser,
     res: express.Response,
     next: express.NextFunction
   ) => {
     try {
+      const role = req.role;
+      if (role != Role.HR) {
+        throw ErrorCodes.UNAUTHORIZED;
+      }
       const employeeDto = plainToInstance(UpdateEmployeeDto, req.body);
       const errors = await validate(employeeDto);
       if (errors.length) {
         console.log(errorsToJson(errors));
         throw new HttpException(400, JSON.stringify(errors));
       }
+
       const employeeId = Number(req.params.id);
       const updateAddress = new UpdateAddressDto();
       if (employeeDto.address) {
         updateAddress.line1 = employeeDto.address.line1;
         updateAddress.pincode = employeeDto.address.pincode;
       }
+
+      const departmentRepository = dataSource.getRepository(Department);
+      const departmentEntity = await departmentRepository.findOneBy({
+        name: employeeDto.department.name,
+      });
+
+      if (!departmentEntity) {
+        throw new EntityNotFoundException({
+          CODE: "DEPARTMENT_NOT_FOUND",
+          MESSAGE: "Department not found",
+        });
+      }
+
       const updatedEmployee = await this.employeeService.updateEmployeeById(
         employeeId,
         employeeDto.name,
         employeeDto.email,
         employeeDto.age,
-        employeeDto.address ? updateAddress : undefined
+        employeeDto.password,
+        employeeDto.role,
+        employeeDto.address ? updateAddress : undefined,
+        departmentEntity
       );
       if (!updatedEmployee) {
-        const error = new HttpException(404, "Employee not found");
-        throw error;
+        throw ErrorCodes.EMPLOYEE_WITH_ID_NOT_FOUND;
       }
       res.status(204).send(updatedEmployee);
     } catch (err) {
@@ -128,20 +160,20 @@ class EmployeeController {
   };
 
   public deleteEmployee = async (
-    req: express.Request,
+    req: RequestWithUser,
     res: express.Response,
     next: express.NextFunction
   ) => {
     try {
+      const role = req.role;
+      if (role != Role.HR) {
+        throw ErrorCodes.UNAUTHORIZED;
+      }
       const employeeId = Number(req.params.id);
       const employee = await this.employeeService.getEmployeeById(employeeId);
 
       if (!employee) {
-        const error = new HttpException(
-          404,
-          `No employee found with id: ${req.params.id}`
-        );
-        throw error;
+        throw ErrorCodes.EMPLOYEE_WITH_ID_NOT_FOUND;
       }
       const deletedEmployee = await this.employeeService.deleteEmployee(
         employeeId
